requirements.txt
proxmoxer
requests
python-dotenv
scikit-learn
numpy
